+++ XCreds/TokenManager.swift
@@ -113,10 +113,7 @@ class TokenManager:DSQueryable {
     static func saveTokensToKeychain(creds:Creds, password:String) -> Bool {
         let keychainUtil = KeychainUtil()
 
-        let passwordItem =  keychainUtil.findPassword(serviceName: "xcreds local password",accountName:PrefKeys.password.rawValue)
-
-        let currentPassword=passwordItem?.password ?? ""
-        if  password.count>0, currentPassword != password, password.count>0 {
+        if  password.count>0 {
             TCSLogWithMark("Saving cloud password")
 
             if keychainUtil.updatePassword(serviceName: "xcreds local password",accountName:PrefKeys.password.rawValue, pass: password, keychainPassword:password) == false {
@@ -125,9 +122,6 @@ class TokenManager:DSQueryable {
                 return false
             }
 
-        }
-        else {
-            TCSLogWithMark("Password update not required")
         }
 
         if let accessToken = creds.accessToken, accessToken.count>0{